Django Framework App
This is a clean copy of the one we have deployed to Heroku. Will use this base with simple pages and DB setup for the rest of project. Will handle deployment later but want to have all the commits and edits in our main github and not in the Heroku system.